sploits/requirements.txt
PyJWT==2.3.0
requests==2.22.0
